refactor: pr 리뷰 내용 반영해서 수정 - 쓰지 않는 코드 주석 삭제 - Object 대신 Pawn을 사용하도록 수정 - 매직넘버 8 -> COL_NUMS로 변경 - Piece 클래스 삭제하고 Pawn 클래스로 통일
